Migrate wallet ledger to integer cents with a canonical balance/heldBalance schema
The wallet's real balance mutations happen across 4 Cloud Functions
(walletFunctions.ts, payhereWebhook.ts, operationalFunctions.ts,
referral.ts) plus the Dart client, and they disagreed on schema before
this change: some tracked a separately-stored availableBalance field,
others only balance/heldBalance, and referral.ts used
FieldValue.increment(double) directly on balance — the exact drift
pattern originally flagged, except live rather than dead code.

Standardize on balance/heldBalance as the only stored fields, with
availableBalance always derived (balance - heldBalance) — every
writer across all 4 functions and the Dart client now agrees, so the
two numbers can no longer disagree. Convert balance, heldBalance,
totalCredited, totalDebited, and every wallet_transactions/
transactions amount/balanceAfter field to integer cents, eliminating
floating-point drift on repeated FieldValue.increment() calls.

Booking.heldAmount/finalPrice/estimatedPrice, Payout amounts, and
referral rewardAmount stay LKR doubles (read/compared elsewhere in
LKR) — every function now converts at the boundary where it reads an
LKR value to credit/debit a wallet, or reads cents back for LKR
display (WalletEntity.availableBalanceLKR, wallet_topup_page.dart,
customer_home_screen.dart, booking_flow_screen.dart's wallet-
sufficiency check).

Fixed two bugs found while rewriting this code: WalletModel never
mapped heldBalance from Firestore at all, so getWallet() always
returned heldBalance: 0 regardless of the real stored value; and
referral.ts logged balanceAfter via FieldValue.increment on a brand
new transaction document, which doesn't represent the wallet's actual
post-credit balance.

Added the wallet repository's first tests (11 cases, including a
top-up-repeated-30-times drift check) and a standalone, dry-run-by-
default migration script (functions/scripts/migrateWalletsToCents.js)
for existing production wallet documents — NOT wired into the deploy
pipeline, NOT executed by this change. See the script's header for
the required backup/sequencing steps before running it for real.

Verified with flutter analyze (0 errors), flutter test (334 passed /
40 failed — same pre-existing failures, confirmed by exact failing-
test-name-set comparison), and `npm run build` (tsc, clean) in
functions/.

+++ lib/features/wallet/data/models/transaction_model.dart
@@ -7,8 +7,8 @@ class TransactionModel {
   final String id;
   final String userId;
   final String type;
-  final double amount;
-  final double balanceAfter;
+  final int amount;
+  final int balanceAfter;
   final String? description;
   final String? relatedBookingId;
   final String? paymentMethod;
@@ -36,8 +36,8 @@ class TransactionModel {
       id: doc.id,
       userId: data['userId'] as String? ?? '',
       type: data['type'] as String? ?? 'topUp',
-      amount: (data['amount'] as num?)?.toDouble() ?? 0.0,
-      balanceAfter: (data['balanceAfter'] as num?)?.toDouble() ?? 0.0,
+      amount: (data['amount'] as num?)?.toInt() ?? 0,
+      balanceAfter: (data['balanceAfter'] as num?)?.toInt() ?? 0,
       description: data['description'] as String?,
       relatedBookingId: data['relatedBookingId'] as String?,
       paymentMethod: data['paymentMethod'] as String?,
